pyproject.toml
[build-system]
requires = ["setuptools>=67.8.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mysql-data-warehouse"
version = "0.1.0"
description = "Data Warehouse project with Python ELT, pandas ETL, PyTorch ML, and analytics artifacts"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
  { name = "Konstantin", email = "your.email@example.com" }
]
requires-python = ">=3.10"

dependencies = [
  "pandas>=2.0.3",
  "python-dotenv>=1.0.0",
  "PyYAML>=6.0",
  "Jinja2>=3.0",
  "torch>=2.1.0",
  "torchvision>=0.15.2",
  "scikit-learn>=1.3.0",
]

[project.optional-dependencies]
dev = [
  "black>=24.3.0",
  "flake8>=6.1.0",
  "pytest>=7.4.2",
  "pytest-mock>=3.15.1",
]

analytics = [
  "tableauhyperapi>=0.0.8422.0",
]

# -----------------------------------------------------
# FIXED BLACK CONFIG (no backslashes, valid TOML)
# -----------------------------------------------------
[tool.black]
line-length = 120
target-version = ["py310", "py311"]
exclude = '''
(
  ^\.venv
 | ^build
 | ^dist
 | ^migrations
 | ^artifacts
 | ^src/runs
 | ^src/templates
)
'''

# -----------------------------------------------------
# FIXED FLAKE8 CONFIG (ignore long lines & artifacts)
# -----------------------------------------------------
[tool.flake8]
max-line-length = 120
extend-ignore = [
  "E501",
  "E203",
  "W503"
]
exclude = [
  ".venv",
  "build",
  "dist",
  "__pycache__",
  "artifacts",
  "src/runs",
  "src/templates",
]

# -----------------------------------------------------
# FIXED ISORT CONFIG
# -----------------------------------------------------
[tool.isort]
profile = "black"
skip = [
  ".venv",
  "build",
  "dist",
  "artifacts",
  "src/runs",
  "src/templates",
]

# -----------------------------------------------------
# PYTEST SECTION — OK
# -----------------------------------------------------
[tool.pytest.ini_options]
minversion = "7.0"
addopts = "--strict-markers --tb=short --durations=10"
testpaths = ["tests"]
markers = [
  "unit: fast, isolated unit tests",
  "contract: data and output contracts for pipeline layers",
  "e2e: end-to-end pipeline tests",
]
